Move toFilterParams to new file within sites/logs/dataviews
